+++ src/main/java/model/order/OrderResponseModel.java
@@ -1,0 +1,22 @@
+package model.order;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import model.order.Order;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderResponseModel {
+
+    private String name;
+
+    private Order order;
+
+    private boolean success;
+
+    private String message;
+}
