--- src/main/java/user/LoggedInUserModel.java
@@ -1,18 +1,0 @@
-package user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoggedInUserModel {
-
-    private User user;
-    private String accessToken;
-    private String refreshToken;
-
-}
