+++ src/test/java/order/make/MakeOrdersStep.java
@@ -1,8 +1,14 @@
 package order.make;
 
+import model.ingredient.IngredientIdListRequestModel;
 import model.ingredient.IngredientListResponseModel;
 import io.qameta.allure.Step;
 import io.restassured.response.Response;
+import model.ingredient.IngredientModel;
+import model.order.OrderResponseModel;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static io.restassured.RestAssured.given;
 
@@ -20,6 +26,26 @@ public class MakeOrdersStep {
     }
 
     @Step
+    public void sendMakeOrderRequest(IngredientIdListRequestModel ingredientIdListRequestModel) {
+
+        response = given()
+                .header("Content-type", "application/json")
+                .and()
+                .body(ingredientIdListRequestModel)
+                .post("api/orders");
+
+    }
+
+    @Step
+    public OrderResponseModel getOrderResponseModel() {
+        OrderResponseModel orderResponseModel = new OrderResponseModel();
+
+        orderResponseModel = response.body().as(OrderResponseModel.class);
+
+        return orderResponseModel;
+    }
+
+    @Step
     public IngredientListResponseModel getIngredientList() {
 
         IngredientListResponseModel ingredientListResponseModel = new IngredientListResponseModel();
@@ -27,5 +53,16 @@ public class MakeOrdersStep {
         ingredientListResponseModel = response.body().as(IngredientListResponseModel.class);
 
         return ingredientListResponseModel;
+    }
+
+    @Step
+    public IngredientIdListRequestModel getIngredientIdListRequestModel() {
+        IngredientIdListRequestModel ingredientIdListRequestModel = new IngredientIdListRequestModel();
+
+        IngredientListResponseModel ingredientListResponseModel = response.body().as(IngredientListResponseModel.class);
+
+        ingredientIdListRequestModel.setIngredients(List.of(ingredientListResponseModel.getRandomBunId(), ingredientListResponseModel.getRandomMainId(), ingredientListResponseModel.getRandomSauceId()));
+
+        return ingredientIdListRequestModel;
     }
 }
