+++ src/test/java/user/create/CreateUserStep.java
@@ -3,8 +3,8 @@ package user.create;
 import io.qameta.allure.Step;
 import io.restassured.response.Response;
 import org.hamcrest.Matchers;
-import user.LoggedInUserModel;
-import user.User;
+import model.user.LoggedInUserModel;
+import model.user.User;
 
 import static io.restassured.RestAssured.given;
 import static org.apache.http.HttpStatus.SC_OK;
